chore: tune golden test
